tree: extract testing module
1. A new module `testing/mod.rs` is created to re-export submodules.
2. `testing.rs` is renamed to `testing/utils.rs`.
3. `integration_testing.rs` is moved to `testing/integration.rs`.
4. `ser_de_tests.rs` is moved to `testing/ser_de_tests.rs`.
5. All imports in the codebase are updated to reflect the new module structure.
