+++ scylla-rust-wrapper/src/testing/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(cpp_integration_testing)]
+pub(crate) mod integration;
+
+#[cfg(test)]
+pub(crate) mod utils;
+
+#[cfg(test)]
+mod ser_de_tests;
